refactor: react-icons를 Icon 컴포넌트로 대체
