+++ src/components/common/GenreButton.tsx
@@ -1,6 +1,6 @@
 'use client';
 import { useState } from 'react';
-import Icon from '../Icons/Icon';
+import Icon from '../../icons/Icon';
 
 interface GenreButtonProps {
   genre: string;
