+++ src/components/common/Header.tsx
@@ -1,20 +1,23 @@
 import Link from 'next/link';
-import Icon from '../Icons/Icon';
+import Icon from '../../icons/Icon';
 import SearchBar from './SearchBar';
-import { RxHamburgerMenu } from 'react-icons/rx';
 
 const Header = () => {
   return (
-    <header className='flex items-center justify-between h-20'>
+    <header className='flex h-20 items-center justify-between'>
       <div className='ml-7 flex w-[25%] gap-10'>
         <button>
-          <RxHamburgerMenu className='w-5 h-5' />
+          <Icon
+            name='HAMBURGER_MENU'
+            size={22}
+            className='stroke-text-sub stroke-[2px]'
+          />
         </button>
 
         <Link href={'/'}>
           <img
             src='/assets/Logo.svg'
-            className='cursor-pointer h-9 w-60'
+            className='h-9 w-60 cursor-pointer'
             alt='Logo'
           />
         </Link>
@@ -26,7 +29,11 @@ const Header = () => {
 
       <div className='mr-7 flex w-[25%] justify-end gap-4'>
         <button>
-          <Icon name='NOTIFICATION' size={23} />
+          <Icon
+            name='NOTIFICATION'
+            size={23}
+            className='stroke-text-sub stroke-[2px]'
+          />
         </button>
         <div>드롭다운</div>
       </div>
